fix(quicklook): 旧 .qlgenerator を Quick Look Preview Extension (.appex) に移行
現行 macOS（Big Sur+/Apple Silicon）は legacy QuickLook Generator
(.qlgenerator) をロードしないため、.mdi のプレビューが一切表示されなかった。
QLPreviewingController ベースの .appex に全面移行する。

- quicklook/MDIQuickLook: GeneratePreviewForURL → PreviewViewController
  (QLPreviewingController + WKWebView)。Markdown 描画は MarkdownRenderer に分離
- Info.plist: BNDL → XPC! / NSExtension(com.apple.quicklook.preview)
- build-quicklook.sh: universal binary を lipo で生成し _NSExtensionMain で
  リンク（旧スクリプトは -target 2 個で x86_64 単独になっていた不具合も解消）
- package.json:
  - extraFiles を Contents/PlugIns/MDIQuickLook.appex に変更
  - mac.extendInfo に UTExportedTypeDeclarations を追加。UTI 未 export だと
    .mdi が dynamic UTI に落ち拡張が永遠にマッチしない根本原因を解消
- electron/quick-look.js と main.js の手動インストール処理を削除
  （.appex は pluginkit が自動検出。旧コードは参照パスも誤っており不動作）
- build/quicklook 出力を .gitignore（CI が build:quicklook で再生成）
