+++ quicklook/MDIQuickLook/Info.plist
@@ -12,50 +12,29 @@
   <string>6.0</string>
   <key>CFBundleName</key>
   <string>MDI Quick Look</string>
+  <key>CFBundleDisplayName</key>
+  <string>MDI Quick Look</string>
   <key>CFBundlePackageType</key>
-  <string>BNDL</string>
+  <string>XPC!</string>
   <key>CFBundleShortVersionString</key>
   <string>1.0</string>
   <key>CFBundleVersion</key>
   <string>1</string>
-  <key>CFBundleDocumentTypes</key>
-  <array>
+  <key>NSExtension</key>
+  <dict>
+    <key>NSExtensionPointIdentifier</key>
+    <string>com.apple.quicklook.preview</string>
+    <key>NSExtensionPrincipalClass</key>
+    <string>MDIQuickLook.PreviewViewController</string>
+    <key>NSExtensionAttributes</key>
     <dict>
-      <key>CFBundleTypeName</key>
-      <string>illusions MDI Document</string>
-      <key>CFBundleTypeRole</key>
-      <string>Viewer</string>
-      <key>LSItemContentTypes</key>
+      <key>QLSupportedContentTypes</key>
       <array>
         <string>com.iktahana.illusions.mdi</string>
       </array>
+      <key>QLSupportsSearchableItems</key>
+      <false/>
     </dict>
-  </array>
-  <key>QLSupportedContentTypes</key>
-  <array>
-    <string>com.iktahana.illusions.mdi</string>
-  </array>
-  <key>UTTypeDeclarations</key>
-  <array>
-    <dict>
-      <key>UTTypeIdentifier</key>
-      <string>com.iktahana.illusions.mdi</string>
-      <key>UTTypeDescription</key>
-      <string>illusions MDI Document</string>
-      <key>UTTypeConformsTo</key>
-      <array>
-        <string>public.plain-text</string>
-      </array>
-      <key>UTTypeTagSpecification</key>
-      <dict>
-        <key>public.filename-extension</key>
-        <array>
-          <string>mdi</string>
-        </array>
-        <key>public.mime-type</key>
-        <string>text/markdown</string>
-      </dict>
-    </dict>
-  </array>
+  </dict>
 </dict>
 </plist>
